requirements.txt
asgiref==3.10.0
asttokens==3.0.0
decorator==5.2.1
Django==5.2.7
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-filter==25.2
django-unfold==0.71.0
executing==2.2.1
factory_boy==3.3.3
Faker==37.12.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
ruff==0.14.4
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.14
